+++ src/oldvexuser.c
@@ -1,0 +1,213 @@
+/*-----------------------------------------------------------------------------*/
+/*                                                                             */
+/*                        Copyright (c) James Pearman                          */
+/*                                   2013                                      */
+/*                            All Rights Reserved                              */
+/*                                                                             */
+/*-----------------------------------------------------------------------------*/
+/*                                                                             */
+/*    Module:     vexuser.c                                                    */
+/*    Author:     James Pearman                                                */
+/*    Created:    7 May 2013                                                   */
+/*                                                                             */
+/*    Revisions:                                                               */
+/*                V1.00  04 July 2013 - Initial release                        */
+/*                                                                             */
+/*-----------------------------------------------------------------------------*/
+/*                                                                             */
+/*    The author is supplying this software for use with the VEX cortex        */
+/*    control system. This file can be freely distributed and teams are        */
+/*    authorized to freely use this program , however, it is requested that    */
+/*    improvements or additions be shared with the Vex community via the vex   */
+/*    forum.  Please acknowledge the work of the authors when appropriate.     */
+/*    Thanks.                                                                  */
+/*                                                                             */
+/*    Licensed under the Apache License, Version 2.0 (the "License");          */
+/*    you may not use this file except in compliance with the License.         */
+/*    You may obtain a copy of the License at                                  */
+/*                                                                             */
+/*      http://www.apache.org/licenses/LICENSE-2.0                             */
+/*                                                                             */
+/*    Unless required by applicable law or agreed to in writing, software      */
+/*    distributed under the License is distributed on an "AS IS" BASIS,        */
+/*    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. */
+/*    See the License for the specific language governing permissions and      */
+/*    limitations under the License.                                           */
+/*                                                                             */
+/*    The author can be contacted on the vex forums as jpearman                */
+/*    or electronic mail using jbpearman_at_mac_dot_com                        */
+/*    Mentor for team 8888 RoboLancers, Pasadena CA.                           */
+/*                                                                             */
+/*-----------------------------------------------------------------------------*/
+
+#include <stdlib.h>
+
+#include "ch.h"          // needs for all ChibiOS programs
+#include "hal.h"         // hardware abstraction layer header
+//#include "vexmotor.h"
+#include "vex.h"        // vex library header
+
+// Digi IO configuration
+static  vexDigiCfg  dConfig[kVexDigital_Num] = {
+        { kVexDigital_1,    kVexSensorDigitalOutput, kVexConfigOutput,      0 },
+        { kVexDigital_2,    kVexSensorDigitalOutput, kVexConfigOutput,      0 },
+        { kVexDigital_3,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_4,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_5,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_6,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_7,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_8,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_9,    kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_10,   kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_11,   kVexSensorDigitalInput,  kVexConfigInput,       0 },
+        { kVexDigital_12,   kVexSensorDigitalInput,  kVexConfigInput,       0 }
+};
+
+static  vexMotorCfg mConfig[kVexMotorNum] = {
+        { kVexMotor_1,      kVexMotor393T,               kVexMotorReversed,       kVexSensorNone,        0 },
+        { kVexMotor_2,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_3,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_4,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_5,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_6,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_7,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_8,      kVexMotorUndefined,          kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_9,      kVexMotor393T,               kVexMotorNormal,       kVexSensorNone,        0 },
+        { kVexMotor_10,     kVexMotor393T,               kVexMotorNormal,       kVexSensorNone,        0 }
+};
+
+
+
+/*-----------------------------------------------------------------------------*/
+/** @brief      User setup                                                     */
+/*-----------------------------------------------------------------------------*/
+/** @details
+ *  The digital and motor ports can (should) be configured here.
+ */
+void
+vexUserSetup()
+{
+    vexDigitalConfigure( dConfig, DIG_CONFIG_SIZE( dConfig ) );
+    vexMotorConfigure( mConfig, MOT_CONFIG_SIZE( mConfig ) );
+}
+
+/*-----------------------------------------------------------------------------*/
+/** @brief      User initialize                                                */
+/*-----------------------------------------------------------------------------*/
+/** @details
+ *  This function is called after all setup is complete and communication has
+ *  been established with the master processor.
+ *  Start other tasks and initialize user variables here
+ */
+#define TRUE  1
+#define FALSE 0
+
+//Ch1 is vertical, Ch2 is horizontal probably
+
+
+void wheelMotion(unsigned char left, unsigned char right) {
+  #define WHEEL_LEFT  kVexMotor_1
+  #define WHEEL_RIGHT kVexMotor_10
+  vexMotorSet( WHEEL_RIGHT, right );
+  vexMotorSet( WHEEL_LEFT , left  );
+}
+
+
+void pollMotion(void) {
+  #define VERT vexControllerGet(Ch1)
+  #define HORI vexControllerGet(Ch2)
+  wheelMotion( VERT - HORI, VERT + HORI );
+}
+
+void vexUserInit(void) {
+  vexAudioPlaySound(256, 100, 100); // say im awake
+
+  while (TRUE) {
+    pollMotion();
+  }
+}
+
+
+/*-----------------------------------------------------------------------------*/
+/** @brief      Autonomous                                                     */
+/*-----------------------------------------------------------------------------*/
+/** @details
+ *  This thread is started when the autonomous period is started
+ */
+msg_t
+vexAutonomous( void *arg )
+{
+    (void)arg;
+
+    // Must call this
+    vexTaskRegister("auton");
+
+    while(0)
+        {
+        // Don't hog cpu
+        //
+        vexSleep( 25 );
+        }
+
+    return (msg_t)0;
+}
+
+#define MotorDriveL     kVexMotor_1
+#define MotorDriveR     kVexMotor_10
+
+/*-----------------------------------------------------------------------------*/
+/** @brief      Driver control                                                 */
+/*-----------------------------------------------------------------------------*/
+/** @details
+ *  This thread is started when the driver control period is started
+ */
+msg_t
+vexOperator( void *arg )
+{
+    int16_t        blink = 0;
+
+    (void)arg;
+
+    // Must call this
+    vexTaskRegister("operator");
+
+    // Run until asked to terminate
+    while(!chThdShouldTerminate())
+        {
+        // flash led/digi out
+
+        // status on LCD of encoder and sonar
+        vexLcdPrintf( VEX_LCD_DISPLAY_2, VEX_LCD_LINE_1, "%4.2fV   %8.1f", vexSpiGetMainBattery() / 1000.0, chTimeNow() / 1000.0 );
+        vexLcdPrintf( VEX_LCD_DISPLAY_2, VEX_LCD_LINE_2, "L %3d R %3d", vexMotorGet( MotorDriveL ), vexMotorGet( MotorDriveR ) );
+
+        // Tank drive
+        // left drive
+        // vexMotorSet( MotorDriveL, vexControllerGet( Ch3 ) );
+
+        // right drive
+        // vexMotorSet( MotorDriveR, vexControllerGet( Ch2 ) );
+
+
+        if (vexControllerGet(Ch1) > 0) {
+            vexMotorSet( MotorDriveL, vexControllerGet(Ch1) );
+        }
+        else {
+            vexMotorSet( MotorDriveL, 0 );
+        }
+
+        if (vexControllerGet(Ch2) > 0) {
+            vexMotorSet( MotorDriveR, vexControllerGet(Ch2) );
+        }
+        else {
+            vexMotorSet( MotorDriveR, 0 );
+        }
+        
+
+        // Don't hog cpu
+        }
+
+    return (msg_t)0;
+}
+
+
+
